Direção do projétil #17

+++ jogador.h
@@ -1,22 +1,19 @@
 //Implementação do jogador, comandos, estruturas etc.
 #ifndef JOGADOR_H
 #define JOGADOR_H
-
 #include "projetil.h"
 
 typedef struct{
 	float x;
 	float y;
 	float velocidade;
-	//Direção do cano do tank.
-	float direcaoCano;
-	//Definição das cores do tank.
+	int direcaoCano;
 	float R;
 	float G;
 	float B;
-	//3 vidas
+	//Quantidade de vidas.
 	float vida;
-	//Vivo ou morto
+	//Vivo ou morto.
 	bool vivo;
 	Projetil projetil;
 }Jogador;
